Исправленные замечания

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -17,7 +17,7 @@ import java.util.HashSet;
 @Data
 public class Film {
 
-    private int id;
+    private long id;
     @NotBlank(message = "Название фильма не может быть пустым")
     private String name;
     @Size(max = 200, message = "Максимальная длина описания - 200 символов")
@@ -26,5 +26,5 @@ public class Film {
     private LocalDate releaseDate;
     @Positive(message = "Продолжительность фильма должна быть положительным числом")
     private int duration;
-    private Set<Integer> likes = new HashSet<>();
+    private Set<Long> likes = new HashSet<>();
 }
